+++ src/paco/application/__init__.py
@@ -16,3 +16,4 @@ from paco.application.reseng_s3bucket import S3BucketResourceEngine
 from paco.application.reseng_eip import EIPResourceEngine
 from paco.application.reseng_ebs import EBSResourceEngine
 from paco.application.reseng_codedeployapplication import CodeDeployApplicationResourceEngine
+from paco.application.reseng_dashboard import DashboardResourceEngine